+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## 1.2.1 (Unreleased)
+
+ENHANCEMENTS:
+
+* resource/linode\_instance: Add `timeouts` support for `create`, `update`, and `delete` (defaults 10, 20, 10)
+* resource/linode\_image: Add `timeouts` support for `create` (defaults 20)
+* resource/linode\_volume: Add `timeouts` support for `create`, `update`, and `delete` (defaults 10, 20, 10)
+
 ## 1.2.0 (November 08, 2018)
 
 ENHANCEMENTS: